Move deprecated pages to own class
