+++ projects/plugins/jetpack/load-jetpack.php
@@ -38,6 +38,7 @@ function jetpack_should_use_minified_assets() {
 add_filter( 'jetpack_should_use_minified_assets', 'jetpack_should_use_minified_assets', 9 );
 
 // @todo: Abstract out the admin functions, and only include them if is_admin()
+require_once JETPACK__PLUGIN_DIR . 'class-jetpack-deprecated.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-network.php';
 require_once JETPACK__PLUGIN_DIR . 'class.jetpack-data.php';
